integration event handler abstraction
